Convert to using new Cog system
Convert the subclassing of the command extension Bot class to making a
Cog instead.  Change launch code to create a Bot instance and add this
cog to it before running the bot.  This is neccessary in order for the
bot to work with the changes to how Cogs works in discord.py 1.0.0
which unfortunately broke the current approach.