Fix is_guild_owner breaking with discord.py 1.5
With discord.py 1.5 the members of a guild is not available unless the
members intent is enabled and used.  This causes the member object of
the guild owner to no longer be cached, which in turn makes the
is_guild_owner check break when trying to fetch the id of it.

Fix by using the owner_id property of the guild which is always
avaliable instead of the owner.